+++ src/RuleDictionnaryPrinter.php
@@ -41,9 +41,6 @@
  **/
 class RuleDictionnaryPrinter extends Rule
 {
-   // From Rule
-    public $can_sort  = true;
-
     public static $rightname = 'rule_dictionnary_printer';